requirements.txt
discord.py
pillow
google-api-python-client
PyNaCl